% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-devtag.R
\name{use_devtag}
\alias{use_devtag}
\title{Use 'devtag' in your package}
\usage{
use_devtag(path = ".")
}
\arguments{
\item{path}{character path to the package}
}
\value{
Nothing: called for file system side effects.
}
\description{
This functions amends DESCRIPTION to set up devtag usage.
Call it once when you're setting a new package.
}
